+++ tests/functional/codegen/calling_convention/test_return_tuple.py
@@ -153,6 +153,50 @@ def test3() -> (address, int128):
     assert c.test3() == [c.out_literals()[2], 1]
 
 
+@pytest.mark.parametrize("string", ["a", "abc", "abcde", "potato"])
+def test_string_inside_tuple(get_contract, string):
+    code = f"""
+@external
+def test_return() -> (String[6], uint256):
+    return "{string}", 42
+    """
+    c1 = get_contract(code)
+
+    code = """
+interface jsonabi:
+    def test_return() -> (String[6], uint256): view
+
+@external
+def test_values(a: address) -> (String[6], uint256):
+    return jsonabi(a).test_return()
+    """
+
+    c2 = get_contract(code)
+    assert c2.test_values(c1.address) == [string, 42]
+
+
+@pytest.mark.parametrize("string", ["a", "abc", "abcde", "potato"])
+def test_bytes_inside_tuple(get_contract, string):
+    code = f"""
+@external
+def test_return() -> (Bytes[6], uint256):
+    return b"{string}", 42
+    """
+    c1 = get_contract(code)
+
+    code = """
+interface jsonabi:
+    def test_return() -> (Bytes[6], uint256): view
+
+@external
+def test_values(a: address) -> (Bytes[6], uint256):
+    return jsonabi(a).test_return()
+    """
+
+    c2 = get_contract(code)
+    assert c2.test_values(c1.address) == [bytes(string, "utf-8"), 42]
+
+
 def test_tuple_return_typecheck(tx_failed, get_contract_with_gas_estimation):
     code = """
 @external
