+++ tests/functional/codegen/types/test_struct.py
@@ -1,6 +1,3 @@
-import pytest
-
-
 def test_nested_struct(get_contract):
     code = """
 struct Animal:
@@ -51,61 +48,3 @@ def modify_nested_single_struct(_human: Human) -> Human:
     c = get_contract(code)
 
     assert c.modify_nested_single_struct({"animal": {"fur": "wool"}}) == (("wool is great",),)
-
-
-@pytest.mark.parametrize("string", ["a", "abc", "abcde", "potato"])
-def test_string_inside_struct(get_contract, string):
-    code = f"""
-struct Person:
-    name: String[6]
-    age: uint256
-
-@external
-def test_return() -> Person:
-    return Person({{ name:"{string}", age:42 }})
-    """
-    c1 = get_contract(code)
-
-    code = """
-struct Person:
-    name: String[6]
-    age: uint256
-
-interface jsonabi:
-    def test_return() -> Person: view
-
-@external
-def test_values(a: address) -> Person:
-    return jsonabi(a).test_return()
-    """
-
-    c2 = get_contract(code)
-    assert c2.test_values(c1.address) == (string, 42)
-
-
-@pytest.mark.parametrize("string", ["a", "abc", "abcde", "potato"])
-def test_string_inside_single_struct(get_contract, string):
-    code = f"""
-struct Person:
-    name: String[6]
-
-@external
-def test_return() -> Person:
-    return Person({{ name:"{string}"}})
-    """
-    c1 = get_contract(code)
-
-    code = """
-struct Person:
-    name: String[6]
-
-interface jsonabi:
-    def test_return() -> Person: view
-
-@external
-def test_values(a: address) -> Person:
-    return jsonabi(a).test_return()
-    """
-
-    c2 = get_contract(code)
-    assert c2.test_values(c1.address) == (string,)
